+++ src/GitCandy.Data.Sqlite/Migrations/GitCandyDbContextModelSnapshot.cs
@@ -53,7 +53,8 @@ namespace GitCandy.Data.Sqlite.Migrations
                     b.HasKey("Id");
 
                     b.HasIndex("NormalizedName")
-                        .IsUnique();
+                        .IsUnique()
+                        .HasDatabaseName("IX_Repositories_NormalizedName");
 
                     b.ToTable("Repositories", (string)null);
                 });
@@ -88,14 +89,17 @@ namespace GitCandy.Data.Sqlite.Migrations
 
                     b.Property<string>("UserId")
                         .IsRequired()
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.HasKey("Id");
 
                     b.HasIndex("Fingerprint")
-                        .IsUnique();
+                        .IsUnique()
+                        .HasDatabaseName("IX_SshKeys_Fingerprint");
 
-                    b.HasIndex("UserId");
+                    b.HasIndex("UserId")
+                        .HasDatabaseName("IX_SshKeys_UserId");
 
                     b.ToTable("SshKeys", (string)null);
                 });
@@ -127,7 +131,8 @@ namespace GitCandy.Data.Sqlite.Migrations
                     b.HasKey("Id");
 
                     b.HasIndex("NormalizedName")
-                        .IsUnique();
+                        .IsUnique()
+                        .HasDatabaseName("IX_Teams_NormalizedName");
 
                     b.ToTable("Teams", (string)null);
                 });
@@ -148,7 +153,8 @@ namespace GitCandy.Data.Sqlite.Migrations
 
                     b.HasKey("TeamId", "RepositoryId");
 
-                    b.HasIndex("RepositoryId");
+                    b.HasIndex("RepositoryId")
+                        .HasDatabaseName("IX_TeamRepositoryRoles_RepositoryId");
 
                     b.ToTable("TeamRepositoryRoles", (string)null);
                 });
@@ -156,6 +162,7 @@ namespace GitCandy.Data.Sqlite.Migrations
             modelBuilder.Entity("GitCandy.Data.Domain.GitCandyUserRepositoryRole", b =>
                 {
                     b.Property<string>("UserId")
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.Property<long>("RepositoryId")
@@ -172,7 +179,8 @@ namespace GitCandy.Data.Sqlite.Migrations
 
                     b.HasKey("UserId", "RepositoryId");
 
-                    b.HasIndex("RepositoryId");
+                    b.HasIndex("RepositoryId")
+                        .HasDatabaseName("IX_UserRepositoryRoles_RepositoryId");
 
                     b.ToTable("UserRepositoryRoles", (string)null);
                 });
@@ -180,6 +188,7 @@ namespace GitCandy.Data.Sqlite.Migrations
             modelBuilder.Entity("GitCandy.Data.Domain.GitCandyUserTeamRole", b =>
                 {
                     b.Property<string>("UserId")
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.Property<long>("TeamId")
@@ -190,7 +199,8 @@ namespace GitCandy.Data.Sqlite.Migrations
 
                     b.HasKey("UserId", "TeamId");
 
-                    b.HasIndex("TeamId");
+                    b.HasIndex("TeamId")
+                        .HasDatabaseName("IX_UserTeamRoles_TeamId");
 
                     b.ToTable("UserTeamRoles", (string)null);
                 });
@@ -198,6 +208,7 @@ namespace GitCandy.Data.Sqlite.Migrations
             modelBuilder.Entity("GitCandy.Data.Identity.GitCandyUser", b =>
                 {
                     b.Property<string>("Id")
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.Property<int>("AccessFailedCount")
@@ -270,6 +281,7 @@ namespace GitCandy.Data.Sqlite.Migrations
             modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityRole", b =>
                 {
                     b.Property<string>("Id")
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("ConcurrencyStamp")
@@ -307,6 +319,7 @@ namespace GitCandy.Data.Sqlite.Migrations
 
                     b.Property<string>("RoleId")
                         .IsRequired()
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.HasKey("Id");
@@ -330,6 +343,7 @@ namespace GitCandy.Data.Sqlite.Migrations
 
                     b.Property<string>("UserId")
                         .IsRequired()
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.HasKey("Id");
@@ -342,9 +356,11 @@ namespace GitCandy.Data.Sqlite.Migrations
             modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserLogin<string>", b =>
                 {
                     b.Property<string>("LoginProvider")
+                        .HasMaxLength(128)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("ProviderKey")
+                        .HasMaxLength(128)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("ProviderDisplayName")
@@ -352,6 +368,7 @@ namespace GitCandy.Data.Sqlite.Migrations
 
                     b.Property<string>("UserId")
                         .IsRequired()
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.HasKey("LoginProvider", "ProviderKey");
@@ -364,9 +381,11 @@ namespace GitCandy.Data.Sqlite.Migrations
             modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserRole<string>", b =>
                 {
                     b.Property<string>("UserId")
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("RoleId")
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.HasKey("UserId", "RoleId");
@@ -379,12 +398,15 @@ namespace GitCandy.Data.Sqlite.Migrations
             modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserToken<string>", b =>
                 {
                     b.Property<string>("UserId")
+                        .HasMaxLength(450)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("LoginProvider")
+                        .HasMaxLength(128)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("Name")
+                        .HasMaxLength(128)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("Value")